working on the fetch problem
