+++ server/app.js
@@ -30,6 +30,9 @@ app.use('/', index);
 app.use('/users', users);
 app.use('/api', api);
 
+var getSummoner = require('./scripts/get_summoner.js');
+console.log(getSummoner);
+
 // catch 404 and forward to error handler
 app.use(function(req, res, next) {
   var err = new Error('Not Found');
